kumiho::pb::kumiho_service_client

Struct KumihoServiceClient

pub struct KumihoServiceClient<T> { /* private fields */ }

The Kumiho service definition.

Implementations

impl KumihoServiceClient<Channel>

pub async fn connect<D>(dst: D) -> Result<Self, Error>
where D: TryInto<Endpoint>, D::Error: Into<StdError>,

Attempt to create a new client by connecting to a given endpoint.

impl<T> KumihoServiceClient<T>
where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F, ) -> KumihoServiceClient<InterceptedService<T, F>>
where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn create_project( &mut self, request: impl IntoRequest<CreateProjectRequest>, ) -> Result<Response<ProjectResponse>, Status>

Project methods

pub async fn get_projects( &mut self, request: impl IntoRequest<GetProjectsRequest>, ) -> Result<Response<GetProjectsResponse>, Status>

pub async fn update_project( &mut self, request: impl IntoRequest<UpdateProjectRequest>, ) -> Result<Response<ProjectResponse>, Status>

pub async fn delete_project( &mut self, request: impl IntoRequest<DeleteProjectRequest>, ) -> Result<Response<StatusResponse>, Status>

pub async fn create_space( &mut self, request: impl IntoRequest<CreateSpaceRequest>, ) -> Result<Response<SpaceResponse>, Status>

Space methods

pub async fn get_space( &mut self, request: impl IntoRequest<GetSpaceRequest>, ) -> Result<Response<SpaceResponse>, Status>

pub async fn get_child_spaces( &mut self, request: impl IntoRequest<GetChildSpacesRequest>, ) -> Result<Response<GetChildSpacesResponse>, Status>

pub async fn delete_space( &mut self, request: impl IntoRequest<DeleteSpaceRequest>, ) -> Result<Response<StatusResponse>, Status>

pub async fn update_space_metadata( &mut self, request: impl IntoRequest<UpdateMetadataRequest>, ) -> Result<Response<SpaceResponse>, Status>

pub async fn create_item( &mut self, request: impl IntoRequest<CreateItemRequest>, ) -> Result<Response<ItemResponse>, Status>

Item methods

pub async fn get_item( &mut self, request: impl IntoRequest<GetItemRequest>, ) -> Result<Response<ItemResponse>, Status>

pub async fn get_items( &mut self, request: impl IntoRequest<GetItemsRequest>, ) -> Result<Response<GetItemsResponse>, Status>

pub async fn item_search( &mut self, request: impl IntoRequest<ItemSearchRequest>, ) -> Result<Response<GetItemsResponse>, Status>

pub async fn delete_item( &mut self, request: impl IntoRequest<DeleteItemRequest>, ) -> Result<Response<StatusResponse>, Status>

pub async fn update_item_metadata( &mut self, request: impl IntoRequest<UpdateMetadataRequest>, ) -> Result<Response<ItemResponse>, Status>

pub async fn search( &mut self, request: impl IntoRequest<SearchRequest>, ) -> Result<Response<SearchResponse>, Status>

Full-text search (Google-like fuzzy search across items)

pub async fn score_revisions( &mut self, request: impl IntoRequest<ScoreRevisionsRequest>, ) -> Result<Response<ScoreRevisionsResponse>, Status>

Score specific revisions against a query (server-side embeddings + fulltext)

pub async fn resolve_kref( &mut self, request: impl IntoRequest<ResolveKrefRequest>, ) -> Result<Response<RevisionResponse>, Status>

Revision methods

pub async fn resolve_location( &mut self, request: impl IntoRequest<ResolveLocationRequest>, ) -> Result<Response<ResolveLocationResponse>, Status>

pub async fn create_revision( &mut self, request: impl IntoRequest<CreateRevisionRequest>, ) -> Result<Response<RevisionResponse>, Status>

pub async fn get_revision( &mut self, request: impl IntoRequest<KrefRequest>, ) -> Result<Response<RevisionResponse>, Status>

pub async fn get_revisions( &mut self, request: impl IntoRequest<GetRevisionsRequest>, ) -> Result<Response<GetRevisionsResponse>, Status>

pub async fn batch_get_revisions( &mut self, request: impl IntoRequest<BatchGetRevisionsRequest>, ) -> Result<Response<BatchGetRevisionsResponse>, Status>

pub async fn delete_revision( &mut self, request: impl IntoRequest<DeleteRevisionRequest>, ) -> Result<Response<StatusResponse>, Status>

pub async fn peek_next_revision( &mut self, request: impl IntoRequest<PeekNextRevisionRequest>, ) -> Result<Response<PeekNextRevisionResponse>, Status>

pub async fn update_revision_metadata( &mut self, request: impl IntoRequest<UpdateMetadataRequest>, ) -> Result<Response<RevisionResponse>, Status>

pub async fn tag_revision( &mut self, request: impl IntoRequest<TagRevisionRequest>, ) -> Result<Response<StatusResponse>, Status>

pub async fn un_tag_revision( &mut self, request: impl IntoRequest<UnTagRevisionRequest>, ) -> Result<Response<StatusResponse>, Status>

pub async fn has_tag( &mut self, request: impl IntoRequest<HasTagRequest>, ) -> Result<Response<HasTagResponse>, Status>

pub async fn was_tagged( &mut self, request: impl IntoRequest<WasTaggedRequest>, ) -> Result<Response<WasTaggedResponse>, Status>

pub async fn set_default_artifact( &mut self, request: impl IntoRequest<SetDefaultArtifactRequest>, ) -> Result<Response<StatusResponse>, Status>

pub async fn create_artifact( &mut self, request: impl IntoRequest<CreateArtifactRequest>, ) -> Result<Response<ArtifactResponse>, Status>

Artifact methods

pub async fn get_artifact( &mut self, request: impl IntoRequest<GetArtifactRequest>, ) -> Result<Response<ArtifactResponse>, Status>

pub async fn get_artifacts( &mut self, request: impl IntoRequest<GetArtifactsRequest>, ) -> Result<Response<GetArtifactsResponse>, Status>

pub async fn get_artifacts_by_location( &mut self, request: impl IntoRequest<GetArtifactsByLocationRequest>, ) -> Result<Response<GetArtifactsByLocationResponse>, Status>

pub async fn delete_artifact( &mut self, request: impl IntoRequest<DeleteArtifactRequest>, ) -> Result<Response<StatusResponse>, Status>

pub async fn update_artifact_metadata( &mut self, request: impl IntoRequest<UpdateMetadataRequest>, ) -> Result<Response<ArtifactResponse>, Status>

pub async fn set_attribute( &mut self, request: impl IntoRequest<SetAttributeRequest>, ) -> Result<Response<StatusResponse>, Status>

Attribute methods (granular metadata operations) These work on any entity type (Revision, Item, Artifact, Space) identified by kref

pub async fn get_attribute( &mut self, request: impl IntoRequest<GetAttributeRequest>, ) -> Result<Response<GetAttributeResponse>, Status>

pub async fn delete_attribute( &mut self, request: impl IntoRequest<DeleteAttributeRequest>, ) -> Result<Response<StatusResponse>, Status>

pub async fn create_edge( &mut self, request: impl IntoRequest<CreateEdgeRequest>, ) -> Result<Response<StatusResponse>, Status>

Edge methods

pub async fn get_edges( &mut self, request: impl IntoRequest<GetEdgesRequest>, ) -> Result<Response<GetEdgesResponse>, Status>

pub async fn delete_edge( &mut self, request: impl IntoRequest<DeleteEdgeRequest>, ) -> Result<Response<StatusResponse>, Status>

pub async fn traverse_edges( &mut self, request: impl IntoRequest<TraverseEdgesRequest>, ) -> Result<Response<TraverseEdgesResponse>, Status>

Graph Traversal methods

pub async fn find_shortest_path( &mut self, request: impl IntoRequest<ShortestPathRequest>, ) -> Result<Response<ShortestPathResponse>, Status>

pub async fn analyze_impact( &mut self, request: impl IntoRequest<ImpactAnalysisRequest>, ) -> Result<Response<ImpactAnalysisResponse>, Status>

pub async fn create_bundle( &mut self, request: impl IntoRequest<CreateBundleRequest>, ) -> Result<Response<ItemResponse>, Status>

Bundle methods

pub async fn add_bundle_member( &mut self, request: impl IntoRequest<AddBundleMemberRequest>, ) -> Result<Response<AddBundleMemberResponse>, Status>

pub async fn remove_bundle_member( &mut self, request: impl IntoRequest<RemoveBundleMemberRequest>, ) -> Result<Response<RemoveBundleMemberResponse>, Status>

pub async fn get_bundle_members( &mut self, request: impl IntoRequest<GetBundleMembersRequest>, ) -> Result<Response<GetBundleMembersResponse>, Status>

pub async fn get_bundle_history( &mut self, request: impl IntoRequest<GetBundleHistoryRequest>, ) -> Result<Response<GetBundleHistoryResponse>, Status>

pub async fn get_tenant_usage( &mut self, request: impl IntoRequest<GetTenantUsageRequest>, ) -> Result<Response<TenantUsageResponse>, Status>

Tenant methods

pub async fn event_stream( &mut self, request: impl IntoRequest<EventStreamRequest>, ) -> Result<Response<Streaming<Event>>, Status>

Event Streaming

pub async fn get_event_capabilities( &mut self, request: impl IntoRequest<GetEventCapabilitiesRequest>, ) -> Result<Response<EventCapabilities>, Status>

Get event streaming capabilities for the authenticated tenant’s tier

pub async fn set_deprecated( &mut self, request: impl IntoRequest<SetDeprecatedRequest>, ) -> Result<Response<StatusResponse>, Status>

Deprecation methods